feat: move the dynamic to the repository domain
